refactor install modal
